requirements.txt
python-dotenv
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
anyio>=3.7.1,<4.0.0
requests
google-auth